Rearranged header files to allow self-contained user includes without requiring BigMatrix (and hence Lapack) stuff. User includes are: SimTKcommon/Scalar.h, SimTKcommon/SmallMatrix.h, SimTKcommon/Orientation.h, and SimTKcommon/Simmatrix.h (with each including the previous ones), and the all-inclusive SimTKcommon.h at the top level.

+++ trunk/Simulation/src/EventHandler.cpp
@@ -87,7 +87,7 @@ PeriodicEventHandler::~PeriodicEventHandler() {
 
 Real PeriodicEventHandler::getNextEventTime(const State& state, bool includeCurrentTime) const {
     Real currentTime = state.getTime();
-    int count = std::floor(currentTime/impl->eventInterval);
+    long count = (long)std::floor(currentTime/impl->eventInterval);
     Real eventTime = count*impl->eventInterval;
     if (eventTime < currentTime || (eventTime == currentTime && !includeCurrentTime)) {
         count++;
